Initial modification of 'process_unknown_words' agent method
Max num of new words integrated into word count is now determined as a random function of lang knowledge and number of days the conversation is repeated ( length ). Number of times a new token is heard is randomly determined via a weighted choice proportional to the relative frequency of each token in conversation.

Word recognition not yet modified.